Testing deploy/resolve settings. Are they in Jenkins or Gradle
Testing changing the deployment repo

Testing branch name/version changes to gradle

Switch the version name based on the branch.

Changes the version to match the branch name.

The default version will be version number with -SNAPSHOT at the end and
the integration branch will deploy to libs-snapshot-local in
Artifactory.

The release version will add -RELEASE to the version number and deploy
to libs-release-local based on the branch being int-release.